requirements.txt
supabase
pandas
plotly
python-dotenv
numpy
